Refactor Model trait CRUD methods
Updated the `update` and `delete` methods to return the result of the query instead of always returning false. Added a check for empty data in `update` to prevent unnecessary queries. Corrected the SQL statement in `delete` to use DELETE instead of UPDATE. Added comments and removed unused code for clarity.